Permission fixed for edit delete and subscriber complete for frontside

+++ app/Http/Controllers/Author/PostController.php
@@ -103,6 +103,11 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
+        if ($post->user_id != Auth::id())
+        {
+            session()->flash('error','You are not authorized to access this post');
+            return redirect()->back();
+        }
         return view('author.post.show',compact('post'));
     }
 
@@ -114,6 +119,11 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+        if ($post->user_id != Auth::id())
+        {
+            session()->flash('error','You are not authorized to access this post');
+            return redirect()->back();
+        }
         $categories = Category::all();
         $tags = Tag::all();
         return view('author.post.edit',compact('post','categories','tags'));
@@ -128,6 +138,11 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        if ($post->user_id != Auth::id())
+        {
+            session()->flash('error','You are not authorized to access this post');
+            return redirect()->back();
+        }
         $this->validate($request,[
             'title' => 'required',
             'image' => 'image',
@@ -190,6 +205,11 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        if ($post->user_id != Auth::id())
+        {
+            session()->flash('error','You are not authorized to access this post');
+            return redirect()->back();
+        }
         if (Storage::disk('public')->exists('post/'.$post->image))
         {
             Storage::disk('public')->delete('post/'.$post->image);
